% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_worm_tensor.R
\name{as_worm_tensor}
\alias{as_worm_tensor}
\title{Generates WormTensor object
A WormTensor object is generated from distance matrices.}
\usage{
as_worm_tensor(Ds)
}
\arguments{
\item{Ds}{A list containing distance matrices}
}
\value{
An object containing distance matrices and metadata
}
\description{
Generates WormTensor object
A WormTensor object is generated from distance matrices.
}
\examples{
\donttest{
    worm_download("mSBD", qc = "PASS")$Ds |> as_worm_tensor() -> object
}
}
