% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worm_membership.R
\name{worm_membership,WormTensor-method}
\alias{worm_membership,WormTensor-method}
\title{Generates membership tensor
A membership tensor is generated from distance matrices.}
\usage{
\S4method{worm_membership}{WormTensor}(object, k)
}
\arguments{
\item{object}{WormTensor object with distance matrices}

\item{k}{Assumed number of clusters}
}
\value{
WormTensor object with membership tensor added
}
\description{
Generates membership tensor
A membership tensor is generated from distance matrices.
}
\examples{
\donttest{
    # Pipe Operation
    worm_download("mSBD", qc = "PASS")$Ds |>
        as_worm_tensor() -> object
    # k=6
    worm_membership(object, k = 6) -> object_k6
}
}
