% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_worm_tensor.R
\name{as_worm_tensor}
\alias{as_worm_tensor}
\title{Generates WormTensor object
A WormTensor object is generated from distance matrices.}
\usage{
as_worm_tensor(Ds)
}
\arguments{
\item{Ds}{A list containing distance matrices}
}
\value{
An object containing distance matrices and metadata
}
\description{
Generates WormTensor object
A WormTensor object is generated from distance matrices.
}
\examples{
#### test####
temp_dl_path <- tempdir()
print(temp_dl_path)
dir.exists(temp_dl_path)

tempfile1 <- file.path(temp_dl_path, "Ds.RData")
download.file("https://figshare.com/ndownloader/files/35963780", tempfile1)

print(tempfile1)
file.exists(tempfile1)
file.access(tempfile1,mode=0)
file.access(tempfile1,mode=1)
file.access(tempfile1,mode=2)
file.access(tempfile1,mode=4)
load(tempfile1)
#### test####
worm_download("mSBD", qc = "PASS")$Ds |> as_worm_tensor() -> object
}
